+++ src/core/entities/user/error.go
@@ -5,3 +5,5 @@ import "errors"
 var ErrIncorrectPassword = errors.New("password incorrect")
 
 var ErrUserNotFound = errors.New("user not found")
+
+var ErrUserAlreadyExists = errors.New("user already exists")