+++ src/core/useCases/create-user-producer/interface.go
@@ -1,0 +1,11 @@
+package create_user_producer
+
+import (
+	dto "go-api/src/presentation/http/controllers/v1/users/create/dto"
+)
+
+type (
+	CreateUserUseCase interface {
+		CreateUser(dto dto.CreateDto) error
+	}
+)