+++ src/main/app.go
@@ -2,8 +2,10 @@ package app
 
 import (
 	"go-api/src/main/container"
-	database "go-api/src/main/module/db"
+	rabbitmq "go-api/src/main/module/amqp/rabbitmq/server"
+	database "go-api/src/main/module/db/mysql"
 	http_server "go-api/src/main/module/http/server"
+	"go-api/src/main/module/work"
 	"log"
 
 	"github.com/joho/godotenv"
@@ -11,8 +13,9 @@ import (
 
 type Server struct {
 	Container *container.Container
-	engine    http_server.HttpServer
+	http      http_server.HttpServer
 	db        database.Database
+	amqp      rabbitmq.RabbitMq
 }
 
 func (server *Server) Setup() *Server {
@@ -20,7 +23,11 @@ func (server *Server) Setup() *Server {
 		container.NewContainerConfig(server.db.Db),
 	)
 
-	server.engine.New(server.Container)
+	work.New(server.Container).StartCrons()
+
+	go server.amqp.New(server.Container).Start()
+
+	server.http.New(server.Container)
 
 	return server
 }