+++ src/main/module/cache/cache.go
@@ -1,0 +1,6 @@
+package cache_client
+
+type CacheClient interface {
+	Set(key string, value string, timeEx int) error
+	Get(key string) (string, error)
+}