+++ src/main/module/http/server/confg-server.go
@@ -3,6 +3,7 @@ package http_server
 import (
 	"go-api/src/main/container"
 	controllers "go-api/src/presentation/http/controllers"
+	v1_user_create "go-api/src/presentation/http/controllers/v1/users/create"
 	v1_user "go-api/src/presentation/http/controllers/v1/users/find-by-di"
 	"go-api/src/presentation/http/middlewares"
 )
@@ -10,6 +11,7 @@ import (
 func loadControllers(container *container.Container) []controllers.Controller {
 	return []controllers.Controller{
 		v1_user.NewController(container),
+		v1_user_create.NewController(container),
 	}
 }