+++ src/presentation/amqp/consumers/users/create/dto/create.dto.go
@@ -1,0 +1,7 @@
+package create_dto
+
+type CreateDto struct {
+	Name     string `json:"name"`
+	Email    string `json:"email"`
+	Password string `json:"password"`
+}