+++ src/presentation/http/controllers/v1/users/find-by-di/find-by-id.go
@@ -8,7 +8,7 @@ import (
 func (createController *createController) findById(req ports_http.HttpRequest) (*ports_http.HttpResponse, *ports_http.HttpResponseError) {
 	id := req.Params.Get("id")
 
-	user, err := createController.container.UserService.GetUser(id)
+	user, err := createController.container.GetUserUseCase.GetUser(id)
 	if err != nil {
 		if err == entity_user.ErrUserNotFound {
 			return nil, &ports_http.HttpResponseError{